# Generate front-end notifications for selected activity on favorite sources