--- a/static/js/components/UpdateProfileForm.jsx
+++ b/static/js/components/UpdateProfileForm.jsx
@@ -21,6 +21,7 @@ import * as ProfileActions from "../ducks/profile";
 
 import UIPreferences from "./UIPreferences";
 import NotificationPreferences from "./NotificationPreferences";
+import FavoriteSourcesNotificationPreferences from "./FavoriteSourcesNotificationPreferences";
 
 const UpdateProfileForm = () => {
   const profile = useSelector((state) => state.profile);
@@ -160,6 +161,9 @@ const UpdateProfileForm = () => {
         </CardContent>
         <CardContent>
           <NotificationPreferences />
+        </CardContent>
+        <CardContent>
+          <FavoriteSourcesNotificationPreferences />
         </CardContent>
         <CardContent>
           <UIPreferences />