This PR introduces an event listener that adds front-end user notifications (when enabled) for activity on users' favorite sources (can be configured to trigger on new comments, spectra and classifications for now -- and can be easily expanded). A generalized SQLAlchemy event listener is used for all three mappers. A new notification settings section is added to the user profile. 

The generated notifications:

![notif](https://user-images.githubusercontent.com/7230285/122281991-c11da700-ce9f-11eb-9089-81c99872d0ea.png)

On the profile page:

![prof](https://user-images.githubusercontent.com/7230285/122282007-c67af180-ce9f-11eb-8a9b-e308406d242f.png)


Closes https://github.com/skyportal/skyportal/issues/1619