@acrellin 2021-06-16
On master: 
``` python
In [12]: DBSession.add(Comment(text="test", author_id=1, obj_id="14gqr"))

In [13]: %time DBSession.commit()                    
CPU times: user 4.81 ms, sys: 0 ns, total: 4.81 ms
Wall time: 40.4 ms
```
On this branch, with 100 test users having added the source in question to their favorites list (probably more than we'll ever encounter in production):
``` python
In [6]: DBSession.add(Comment(text="test", author_id=1, obj_id="14gqr"))

In [7]: %time DBSession.commit()
CPU times: user 371 ms, sys: 19.8 ms, total: 391 ms
Wall time: 468 ms
```
So yeah, definitely slows down the inserts a bit, but as these aren't resources that will be programmatically added in large numbers, the performance seems reasonable to me here, especially considering it's hard to imagine 100 users all having the same source on their favorites list and all having these types of notifications enabled (but even if so, this performance still seems acceptable considering the intermittent nature & low frequency of adding spectra, comments, and classifications). Thoughts @kmshin1397 ?