@kmshin1397 2021-06-16
Hmm, I would agree with that when it comes to comments and classifications. But my impressions is that spectra are often added programmatically within a short time frame, at least when it comes to SEDM spectra. I think @scizen9 has scripts which do that sort of thing, if I recall correctly. Is that true @scizen9? You upload spectra through scripts, right? Do those tend to be a lot of spectra at once? How would you feel about a increase in time per spectrum post?