@acrellin 2021-06-16
@kmshin1397 @scizen9 As the overhead here is due to inserting new `UserNotification` rows, the overhead should be a linearly-scaling ~400ms per 100 users who are watching the source in question and have enabled the relevant notification types. Does that sound right?