@acrellin 2021-06-17
Here's with 100 users watching the obj in question, but without having enabled the relevant notifications: 
``` python
In [7]: %time DBSession.commit()                    
CPU times: user 11.1 ms, sys: 614 µs, total: 11.7 ms
Wall time: 14.2 ms
```