@acrellin 2021-06-17
Updated notification text styling: 
![notif](https://user-images.githubusercontent.com/7230285/122310550-6c8d2280-cec5-11eb-8972-7bb46901a606.png)