@guynir42 2021-06-17
This looks good overall. Can you remind me what is the difference between "mark as read" and "delete" in this context?