@acrellin 2021-06-17
> This looks good overall. Can you remind me what is the difference between "mark as read" and "delete" in this context?

That's just the preexisting functionality for the browser notifications - they can be marked read / unread and deleted.