@kmshin1397 2021-06-17
Note: approved since while spectra are posted programmatically, they're done so fairly sparsely so the performance hit shouldn't really matter